# configs/swagger/v1.yaml
swagger: "2.0"
info:
  description: "悟空聊天IM API"
  version: "1.0.0"
  title: "悟空聊天IM API"
host: "49.235.59.182:8080"
...
schemes:
  - "http"
basePath: "/v1"
paths:
...
  /user/search:
    get:
      tags:
        - "user"
      summary: "搜索用户（精准查询）"
      description: "搜索用户（精准查询"
      operationId: "userSearch"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "keyword"
          type: string
          description: "关键字"
          required: true
      responses:
        200:
          description: "返回"
          schema:
            $ref: '#/definitions/UserExistResp'
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: []      
  /users/{uid}:
    get:
      tags:
        - "user"
      summary: "获取指定uid的用户信息"
      description: "获取指定uid的用户信息"
      operationId: "userGet"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "uid"
          type: string
          description: "用户uid"
          required: true
      responses:
        200:
          description: "返回"
          schema:
            $ref: '#/definitions/UserDetailResp'
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response' 
      security:
        - token: []      
...
  /friend/apply:
    post:
      tags:
        - "friend"
      summary: "申请加好友"
      description: "申请加好友,系统将发送消息给被加人，消息里带有这次操作的凭证token"
      operationId: "friendApply"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "req"
          description: "请求数据"
          required: true
          schema:
            $ref: "#/definitions/ApplyReq"
      responses:
        200:
          description: "成功"
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: []
  /friend/sure:
    post:
      tags:
        - "friend"
      summary: "确认好友申请"
      description: "确认好友申请（接收好友申请的人，确认当前申请的时候调用，当确认成功后系统将发送消息给对方，对方收到消息后需要同步自己的好友数据）"
      operationId: "friendSure"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "req"
          description: "请求数据"
          required: true
          schema:
            $ref: "#/definitions/SureReq"
      responses:
        200:
          description: "成功"
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: []
  /friend/sync:
    post:
      tags:
        - "friend"
      summary: "同步好友信息"
      description: "同步好友信息"
      operationId: "friendSync"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "sync_key"
          type: "string"
          description: "同步key,格式： 最新更新时间@id"
        - in: "query"
          name: "limit"
          type: integer
          format: int64
          description: "最大限制"
      responses:
        200:
          description: "成功"
          schema:
            type: array
            items:
              $ref: '#/definitions/FriendResp'
        400:
          description: '错误'
          schema:
            $ref: '#/definitions/response'
      security:
        - token: []
...
securityDefinitions:
  token:
    type: "apiKey"
    in: "header"
    name: "token"
    description: "用户token"
definitions:
...
  UserDetailResp:
    type: "object"
    properties:
      uid:
        type: string
        description: "用户唯一ID"
      name:
        type: string
        description: "用户名字" 
      mute:
        type: integer
        format: int
        description: "免打扰"
      top:  
        type: integer
        format: int
        description: "置顶"  
...
  UserExistResp:
    type: "object"
    properties:
      exist:
        type: integer
        format: int
        description: "是否存在用户 0.不存在 1.存在"
      data:
        $ref: '#/definitions/UserResp'
  UserResp:
    type: "object"
    properties:
      uid:
        type: string
        description: "用户唯一ID"
      name:
        type: string
        description: "用户名字"
  FriendResp:
    type: "object"
    properties:
      id:
        type: integer
        format: int64
        description: "唯一id,增量同步会用到"
      to_uid:
        type: string
        description: "好友uid"
      to_name:
        type: string
        description: "好友名称"
      created_at:
        type: string
        description: "创建时间"
      updated_at:
        type: string
        description: "更新时间"
  SureReq:
    type: "object"
    properties:
      token:
        type: "string"
        description: "申请加好友的凭证"
  ApplyReq:
    type: "object"
    properties:
      to_uid:
        type: "string"
        description: "向谁申请好友"
      remark:
        type: "string"
        description: "加好友的备注"
...
  response:
    type: "object"
    properties:
      status:
        type: integer
        format: int
      msg:
        type: "string"